Update to rustix 0.38.
This also splits all `extern "C"` functions out of c-scape/src/lib.rs, to fix
compiler warnings.
