+++ c-gull/Cargo.toml
@@ -20,7 +20,7 @@ tz-rs = { version = "0.6.11", optional = true }
 printf-compat = { version = "0.1.1", optional = true }
 sync-resolve = { version = "0.3.0", optional = true }
 log = { version = "0.4.14", default-features = false, optional = true }
-rustix = { version = "0.37.17", default-features = false, optional = true, features = ["fs", "itoa", "net", "param", "process", "procfs", "rand", "termios", "thread", "time"] }
+rustix = { version = "0.38.1", default-features = false, optional = true, features = ["fs", "itoa", "net", "param", "process", "procfs", "rand", "termios", "thread", "time"] }
 
 [dev-dependencies]
 libc = "0.2.138"
