+++ c-scape/src/io/pipe.rs
@@ -1,5 +1,5 @@
 use rustix::fd::IntoRawFd;
-use rustix::io::PipeFlags;
+use rustix::pipe::PipeFlags;
 
 use libc::c_int;
 
@@ -9,7 +9,7 @@ use crate::convert_res;
 unsafe extern "C" fn pipe(pipefd: *mut c_int) -> c_int {
     libc!(libc::pipe(pipefd));
 
-    match convert_res(rustix::io::pipe()) {
+    match convert_res(rustix::pipe::pipe()) {
         Some((a, b)) => {
             *pipefd = a.into_raw_fd();
             *pipefd.add(1) = b.into_raw_fd();
@@ -25,7 +25,7 @@ unsafe extern "C" fn pipe2(pipefd: *mut c_int, flags: c_int) -> c_int {
     libc!(libc::pipe2(pipefd, flags));
 
     let flags = PipeFlags::from_bits(flags as _).unwrap();
-    match convert_res(rustix::io::pipe_with(flags)) {
+    match convert_res(rustix::pipe::pipe_with(flags)) {
         Some((a, b)) => {
             *pipefd = a.into_raw_fd();
             *pipefd.add(1) = b.into_raw_fd();
