+++ c-scape/src/io/poll.rs
@@ -8,10 +8,10 @@ use crate::convert_res;
 unsafe extern "C" fn poll(fds: *mut libc::pollfd, nfds: libc::nfds_t, timeout: c_int) -> c_int {
     libc!(libc::poll(fds, nfds, timeout));
 
-    let pollfds: *mut rustix::io::PollFd = checked_cast!(fds);
+    let pollfds: *mut rustix::event::PollFd = checked_cast!(fds);
 
     let fds = slice::from_raw_parts_mut(pollfds, nfds.try_into().unwrap());
-    match convert_res(rustix::io::poll(fds, timeout)) {
+    match convert_res(rustix::event::poll(fds, timeout)) {
         Some(num) => num.try_into().unwrap(),
         None => -1,
     }
