+++ c-scape/src/io/splice.rs
@@ -1,5 +1,5 @@
 use rustix::fd::BorrowedFd;
-use rustix::io::SpliceFlags;
+use rustix::pipe::SpliceFlags;
 
 use libc::{c_int, c_uint, loff_t};
 
@@ -20,7 +20,7 @@ unsafe extern "C" fn splice(
     let off_in: *mut u64 = checked_cast!(off_in);
     let off_out: *mut u64 = checked_cast!(off_out);
 
-    match convert_res(rustix::io::splice(
+    match convert_res(rustix::pipe::splice(
         BorrowedFd::borrow_raw(fd_in),
         off_in.as_mut(),
         BorrowedFd::borrow_raw(fd_out),
@@ -43,9 +43,9 @@ unsafe extern "C" fn vmsplice(
 ) -> libc::ssize_t {
     libc!(libc::vmsplice(fd, iov, nr_segs, flags));
 
-    let bufs = core::slice::from_raw_parts(iov.cast::<rustix::io::IoSliceRaw>(), nr_segs);
+    let bufs = core::slice::from_raw_parts(iov.cast::<rustix::pipe::IoSliceRaw>(), nr_segs);
     let flags = SpliceFlags::from_bits(flags).unwrap();
-    match convert_res(rustix::io::vmsplice(
+    match convert_res(rustix::pipe::vmsplice(
         BorrowedFd::borrow_raw(fd),
         bufs,
         flags,
