+++ c-scape/src/process/priority.rs
@@ -1,6 +1,6 @@
 use crate::convert_res;
 use errno::{set_errno, Errno};
-use libc::{c_int, c_uint, id_t};
+use libc::{c_int, c_uint, id_t, pid_t};
 use rustix::process::{Pid, Uid};
 
 #[no_mangle]
@@ -8,8 +8,8 @@ unsafe extern "C" fn getpriority(which: c_uint, who: id_t) -> c_int {
     libc!(libc::getpriority(which, who));
 
     let result = match which {
-        libc::PRIO_PROCESS => rustix::process::getpriority_process(Pid::from_raw(who)),
-        libc::PRIO_PGRP => rustix::process::getpriority_pgrp(Pid::from_raw(who)),
+        libc::PRIO_PROCESS => rustix::process::getpriority_process(Pid::from_raw(who as pid_t)),
+        libc::PRIO_PGRP => rustix::process::getpriority_pgrp(Pid::from_raw(who as pid_t)),
         libc::PRIO_USER => rustix::process::getpriority_user(Uid::from_raw(who)),
         _ => {
             set_errno(Errno(libc::EINVAL));
@@ -28,8 +28,10 @@ unsafe extern "C" fn setpriority(which: c_uint, who: id_t, prio: c_int) -> c_int
     libc!(libc::setpriority(which, who, prio));
 
     let result = match which {
-        libc::PRIO_PROCESS => rustix::process::setpriority_process(Pid::from_raw(who), prio),
-        libc::PRIO_PGRP => rustix::process::setpriority_pgrp(Pid::from_raw(who), prio),
+        libc::PRIO_PROCESS => {
+            rustix::process::setpriority_process(Pid::from_raw(who as pid_t), prio)
+        }
+        libc::PRIO_PGRP => rustix::process::setpriority_pgrp(Pid::from_raw(who as pid_t), prio),
         libc::PRIO_USER => rustix::process::setpriority_user(Uid::from_raw(who), prio),
         _ => {
             set_errno(Errno(libc::EINVAL));
