+++ origin/Cargo.toml
@@ -11,8 +11,8 @@ repository = "https://github.com/sunfishcode/mustang"
 edition = "2021"
 
 [dependencies]
-linux-raw-sys = { version = "0.3.4", default-features = false, features = ["general", "no_std"] }
-rustix = { version = "0.37.17", default-features = false, features = ["mm", "thread", "time", "runtime", "param", "process"] }
+linux-raw-sys = { version = "0.4.3", default-features = false, features = ["general", "no_std"] }
+rustix = { version = "0.38.1", default-features = false, features = ["mm", "thread", "time", "runtime", "param", "process"] }
 bitflags = "1.3.0"
 memoffset = { version = "0.8.0", optional = true }
 log = { version = "0.4.14", default-features = false, optional = true }
